feat(main.ts): introduce getEmbeddingModel utility function for embedding model management feat(main.ts): update SmartSeekerPlugin to use getEmbeddingModel for embedding model instantiation
